requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
jsonlines==3.1.0
lmdb==1.3.0
numpy==1.19.2
outcome==1.2.0
pathlib==1.0.1
Pillow==9.2.0
PySocks==1.7.1
PyVirtualDisplay==3.0
requests==2.28.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
timm==0.6.7
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.64.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.12
wsproto==1.2.0
zhon==1.1.5
